requirements.txt
--extra-index-url=https://pypi.nvidia.com
cupy-cuda12x
cudf-cu12==24.4.*
dask-cudf-cu12==24.4.*
cuml-cu12==24.4.*
cugraph-cu12==24.4.*
cuspatial-cu12==24.4.*
cuproj-cu12==24.4.*
cuxfilter-cu12==24.4.*
cucim-cu12==24.4.*
pylibraft-cu12==24.4.*
raft-dask-cu12==24.4.*
cuvs-cu12==24.4.*
torch
pandas
tqdm
scikit-learn
h5py
transformers
jsonlines
lightning
protobuf==3.20
xxhash
wandb
sentencepiece
tensorboardX
datasets
huggingface_hub[cli]
pyarrow